requirements.txt
# --- Core Framework & Utilities ---
gradio>=4.0.0
fastapi
uvicorn[standard]
python-multipart

pyyaml
piexif
pillow
tqdm
huggingface_hub>=0.24.3

# --- Key Libraries ---
# Install torch manually
# transformers 5.x required for qwen3_5 architecture (Qwen3.5-9B models).
# If pip can't resolve >=5.0.0 from PyPI yet, install manually first:
#   pip install git+https://github.com/huggingface/transformers.git
transformers>=5.0.0
accelerate>=0.26.0
safetensors

# --- Shared Model Dependencies ---
timm
einops
sentencepiece
protobuf
bitsandbytes
scipy

# --- Model Specific Dependencies ---

# ToriiGate / Qwen2-VL
qwen_vl_utils
# flash-attn --no-build-isolation  # Qwen3-VL: Flash Attention 2 support (manual install required)
# PyQt5 (optional, used by some ToriiGate variants)

# JoyCaption / Llama
# liger_kernel # Linux only (Triton dependency)

# Pixtral / JoyTag
tiktoken

# SmolVLM2
num2words

# Miscellaneous
# spaces # HuggingFace Spaces specific helper
colorama

# WD14 Tagger
onnxruntime-gpu
opencv-python-headless